Store damage as an integer

+++ src/game/objects/har.h
@@ -143,7 +143,7 @@ typedef struct har_t {
     uint8_t air_attacked;
     uint8_t is_wallhugging;          // HAR is standing right next to a wall
     uint8_t is_grabbed;              // Is being moved by another object. Set by ex, ey tags
-    float last_damage_value;         // Last damage value taken
+    int16_t last_damage_value;       // Last damage value taken
     har_endurance_t last_stun_value; // Last stun value taken. HAR_ENDURANCE fixed point.
 
     float jump_speed;      // Agility generated speed modifier for jumping
